project_directory/app/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
torch==2.0.0+cpu
torchvision==0.15.0+cpu
transformers==4.33.0
facenet-pytorch==2.5.2
safetensors==0.3.1
pandas==2.0.3
tqdm==4.65.0
pillow==9.5.0
numpy==1.22.0
safetensors==0.3.1
python-multipart==0.0.6